Move Loader component to new folder + update import
